weight

+++ libraries/ScaleModule/ScaleModule.h
@@ -9,7 +9,7 @@ class ScaleModule {
     ScaleModule(uint8_t doutPin, uint8_t sckPin);
     void begin();
     void tare();         // Zero the scale manually
-    long read();         // Get raw value
+    float read();         // Get raw value
     bool isReady();      // Check if HX711 is ready
 
   private:
